back-end/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
geographiclib==1.50
geopy==2.1.0
idna==2.10
mysqlclient==1.4.6
numpy==1.20.1
pip==21.0.1
PyMySQLv1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
setuptools==54.1.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.58.0
unicodecsv==0.14.1
urllib3==1.26.3
wheel==0.36.2
Werkzeug==1.0.1
pyOpenSSL==20.0.1
django-extensions==3.1.1
